src/main/java/com/arthurprojects/vcs/service/RepositoryService.java
package com.arthurprojects.vcs.service;


import com.arthurprojects.vcs.model.Repository;
import org.springframework.stereotype.Service;

import java.io.IOException;
import java.nio.file.Path;

@Service
public class RepositoryService {

    private static final String VCS_DIR = ".vcs";

    public Repository initRepository(Path path) throws IOException {
        Path vcsDir = path.resolve(VCS_DIR);
        if (!vcsDir.toFile().exists()) {
            vcsDir.toFile().mkdir();
            // Create necessary subdirectories
            createVcsStructure(vcsDir);
        }
        Repository repo = new Repository();
        // Initialize repository structure
        return repo;
    }

    private void createVcsStructure(Path vcsDir) throws IOException {
        // Create objects directory for storing blobs
        Path objectsDir = vcsDir.resolve("objects");
        objectsDir.toFile().mkdir();

        // Create refs directory for branches
        Path refsDir = vcsDir.resolve("refs");
        refsDir.toFile().mkdir();

        // Create initial HEAD file pointing to master branch
        Path headFile = vcsDir.resolve("HEAD");
        // Initialize HEAD file
    }

    public Repository cloneRepository(Path source, Path destination) throws IOException {
        // Implement repository cloning logic
        return null;
    }
}
